+++ apps/relay/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @farcaster/auth-relay
 
+## 0.2.1
+
+### Patch Changes
+
+- 827053c: drop SIWE params in constructUrl
+
+## 0.2.0
+
+### Minor Changes
+
+- 088f6ca: Drop SIWE params from sign in URLs
+
 ## 0.1.1
 
 ### Patch Changes